Sync w. latest menus - Photon_Cleaned, Ecal_Activity, CSC_Activity, L1MuOpen_DT bits.

+++ HLTrigger/HLTanalyzers/test/RateEff/OHltMenu.cpp
@@ -11,7 +11,7 @@ void OHltMenu::SetMapL1SeedsOfStandardHLTPath(std::map<TString, std::vector<TStr
   //std::cout<<map_L1SeedsOfStandardHLTPath.size()<<std::endl;
 }
 
-void OHltMenu::AddTrigger(TString trign, float presc, float eventS) {
+void OHltMenu::AddTrigger(TString trign, int presc, float eventS) {
   //used for L1 menu developemnt
   names.push_back(trign);
   seedcondition[trign] 	       	= trign;  // l1 is it's own seed
@@ -19,7 +19,7 @@ void OHltMenu::AddTrigger(TString trign, float presc, float eventS) {
   eventSizes[trign] 	       	= eventS;
 }
 
-void OHltMenu::AddTrigger(TString trign, TString seedcond, float presc, float eventS, int refpresc) {
+void OHltMenu::AddTrigger(TString trign, TString seedcond, int presc, float eventS, int refpresc) {
   names.push_back(trign);
   seedcondition[trign] 	       	= seedcond;
   prescales[trign] 	       	= presc;
@@ -27,7 +27,7 @@ void OHltMenu::AddTrigger(TString trign, TString seedcond, float presc, float ev
   referenceRunPrescales[trign]  = refpresc;
 }
 
-void OHltMenu::AddL1forPreLoop(TString trign, float presc) {
+void OHltMenu::AddL1forPreLoop(TString trign, int presc) {
   L1names.push_back(trign);
   L1prescales[trign] 	       	= presc;
 }
